refactor(types): Split out our types module
